LayoutTests/imported/w3c/web-platform-tests/dom/collections/HTMLCollection-own-props-expected.txt


FAIL Setting non-array index while named property exists (loose) assert_equals: expected (object) Element node <a id="named"></a> but got (string) "foo"
FAIL Setting non-array index while named property exists (strict) assert_throws_js: function "function () {
    c[name] = "foo";
  }" did not throw
PASS Setting non-array index while named property doesn't exist (loose) 
PASS Setting non-array index while named property doesn't exist (strict) 
PASS Setting array index while indexed property exists (loose) 
FAIL Setting array index while indexed property exists (strict) assert_throws_js: function "function () {
    c[0] = "foo";
  }" did not throw
FAIL Setting array index while indexed property doesn't exist (loose) assert_equals: expected (undefined) undefined but got (string) "foo"
FAIL Setting array index while indexed property doesn't exist (strict) assert_equals: expected (undefined) undefined but got (string) "foo"
